# api/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Criação de uma API com nodeJS, para disponibilizar dados do livro de Monstros do D&D 5e, esse projeto é um trabalho em grupo da materia Arquitetura de computadores, que visa colaborar com a comunidade de RPG de mesa. https://github.com/laurencioScript/api_monster",
    "version": "1.0.0",
    "title": "API Monster D&D"
  },
  "host": "0.0.0.0:8080",
...
  "paths": {
...
    "/monster/getPericia/": {
      "get": {
        "tags": [
          "Monstro"
        ],
        "summary": "",
        "description": "Busca todos os monstros de acordo com o parametros passados, não deixe espaço vazio substitua por null.",
        "operationId": "getPericia",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "parm1",
            "in": "query",
            "description": "pericia do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "parm2",
            "in": "query",
            "description": "pericia do monstro",
            "required": false,
            "type": "string"
          },
          {
            "name": "parm3",
            "in": "query",
            "description": "pericia do monstro",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Monstros": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "nome": {
                        "type": "string"
                      },
                      "pericia": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "tipo": {
                              "type": "string"
                            },
                            "bonus": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
...
    "/analise/getAllTipo/": {
      "get": {
        "tags": [
          "Dados"
        ],
        "summary": "",
        "description": "",
        "operationId": "getAllTipo",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Tipos": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Tipo": {
                        "type": "string"
                      },
                      "Quantidade": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
    "/analise/getAllSentidos/": {
      "get": {
        "tags": [
          "Dados"
        ],
        "summary": "",
        "description": "",
        "operationId": "getAllSentidos",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Sentidos": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Sentido": {
                        "type": "string"
                      },
                      "Quantidade": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
...
    "/analise/getAllImunidades/": {
      "get": {
        "tags": [
          "Dados"
        ],
        "summary": "",
        "description": "",
        "operationId": "getAllImunidades",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Imunidades": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Imunidade": {
                        "type": "string"
                      },
                      "Quantidade": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
...
    "/analise/getAllReacoes/": {
      "get": {
        "tags": [
          "Dados"
        ],
        "summary": "",
        "description": "",
        "operationId": "getAllReacoes",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Reacoes": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Reacao": {
                        "type": "string"
                      },
                      "Quantidade": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
    "/analise/getAllEquipamentos/": {
      "get": {
        "tags": [
          "Dados"
        ],
        "summary": "",
        "description": "",
        "operationId": "getAllEquipamentos",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Equipamentos": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "Equipamento": {
                        "type": "string"
                      },
                      "Quantidade": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    },
    "/analise/getAllResistenciaDano/": {
      "get": {
        "tags": [
          "Dados"
        ],
        "summary": "",
        "description": "",
        "operationId": "getAllResistenciaDano",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso na operação",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer"
                },
                "Resistencias de Dano": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "ResistenciaDano": {
                        "type": "string"
                      },
                      "Quantidade": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ta bugado me chama laurencio.arkauss@gmail.com "
          },
          "404": {
            "description": "Ausência de parâmetros"
          }
        }
      }
    }
  },
...
}